Move all the rendering logic to a dedicated file
